+++ versions/1.21.5/src/main/java/at/hannibal2/skyhanni/mixins/hooks/MessageHandlerHook.kt
@@ -108,6 +108,10 @@ private fun handleOnGameMessage(message: Text, actionBar: Boolean, original: BiC
     }
     val (result, cancel) = ChatManager.onChatReceive(message)
     result?.let {
+        // Will send both the unmodified and modified message into the Fabric Pipeline so other mods also get the old unmodified message
+        // This sadly isn't preventable without switching fully to Fabric Chat Events
+        // (which needs an event for cancelling and an event for modifying, which isn't a feasible split up with this code base size)
+        ClientReceiveMessageEvents.ALLOW_GAME.invoker().allowReceiveGameMessage(message, actionBar)
         original.accept(it, actionBar)
         return
     }